feat: [api-handlers-decomposition-recovery-003] - Extract factory and session route handlers
